Controller acabado

+++ src/main/java/com/eduardobenavides/service/TeamService.java
@@ -17,25 +17,25 @@ public class TeamService {
     private TeamRepository teamRepository;
 
     public void createTeams(){
-        Team team1 = new Team(1,"Cavaliers","Cleveland", LocalDate.of(1970,10,12));
+        Team team1 = new Team((long) 1,"Cavaliers","Cleveland", LocalDate.of(1970,10,12));
         teamRepository.save(team1);
 
-        Team team2 = new Team(2,"Warriors","Oakland", LocalDate.of(1946,4,25));
+        Team team2 = new Team((long) 2,"Warriors","Oakland", LocalDate.of(1946,4,25));
         teamRepository.save(team2);
 
-        Team team3 = new Team(3,"Pelicans","New Orleans", LocalDate.of(2013,1,4));
+        Team team3 = new Team((long) 3,"Pelicans","New Orleans", LocalDate.of(2013,1,4));
         teamRepository.save(team3);
 
-        Team team4 = new Team(4,"Rockets","Houston", LocalDate.of(1967,10,30));
+        Team team4 = new Team((long) 4,"Rockets","Houston", LocalDate.of(1967,10,30));
         teamRepository.save(team4);
 
-        Team team5 = new Team(5,"Tunder","Oklahoma", LocalDate.of(1979,6,7));
+        Team team5 = new Team((long) 5,"Tunder","Oklahoma", LocalDate.of(1979,6,7));
         teamRepository.save(team5);
 
-        Team team6 = new Team(6,"Clippers","Los Angeles",LocalDate.of(1970,7,19));
+        Team team6 = new Team((long) 6,"Clippers","Los Angeles",LocalDate.of(1970,7,19));
         teamRepository.save(team6);
 
-        Team team7 = new Team(7,"Lakers","Los Angeles",LocalDate.of(1946,8,19));
+        Team team7 = new Team((long) 7,"Lakers","Los Angeles",LocalDate.of(1946,8,19));
         teamRepository.save(team7);
 
     }
